SN_BOT_NEW/swedish_municipal_crawler/requirements.txt
scrapy>=2.8.0
scrapy-playwright>=0.0.25
playwright>=1.30.0
camelot-py[cv]>=0.10.1
pdfplumber>=0.9.0
opencv-python>=4.7.0
pandas>=1.5.0
openpyxl>=3.1.0
redis>=4.5.0
requests>=2.28.0
lxml>=4.9.0
cssselect>=1.2.0
itemadapter>=0.7.0
beautifulsoup4==4.12.2
flask==3.0.0
flask-socketio>=5.3.6
sqlalchemy==2.0.23
ghostscript==0.7
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
nest-asyncio>=1.5.0
psutil>=5.9.6 